Moved 'positioned_tap_detector_2.dart' to more appropriate directory
