+++ lib/flutter_map.dart
@@ -54,4 +54,3 @@ export 'package:flutter_map/src/misc/fit_bounds_options.dart';
 export 'package:flutter_map/src/misc/move_and_rotate_result.dart';
 export 'package:flutter_map/src/misc/point_extensions.dart';
 export 'package:flutter_map/src/misc/position.dart';
-export 'package:flutter_map/src/misc/private/positioned_tap_detector_2.dart';
